requirements.txt
pika
requests
tenacity
python-dotenv